Moved SizedQueue tto util folder
